requirements.txt
django
django-crispy-forms
Pillow
djangorestframework